updated unit tests after refactor to SubmissionDirectoryService
